+++ packages/signals/lib/signals.dart
@@ -22,7 +22,7 @@ export 'src/signals.dart'
         ReadonlySignal;
 export 'src/utils/timer.dart' show TimerSignal, TimerExtension;
 export 'src/value/async_signal.dart'
-    show futureSignalWithDefault, streamSignalWithDefault, AsyncSignal;
+    show asyncSignalFromFuture, asyncSignalFromStream, AsyncSignal;
 export 'src/value/future_signal.dart' show futureSignal, FutureSignal;
 export 'src/value/iterable_signal.dart' show iterableSignal, IterableSignal;
 export 'src/value/list_signal.dart' show listSignal, ListSignal;