+++ packages/signals/lib/src/signals.dart
@@ -436,6 +436,15 @@ typedef MutableSignal<T> = Signal<T>;
 /// Signal that can be extended and used as a class
 class ValueSignal<T> extends _Signal<T> {
   ValueSignal(super.value, {super.debugLabel});
+
+  /// Should only be called to update the value of a signal if checks for equality
+  /// have already been made.
+  ///
+  /// This is primarily used by the `ValueSignal` class to update the value
+  /// since the reference is always the same.
+  void forceUpdate(T val) {
+    _updateValue(val);
+  }
 }
 
 class _Signal<T> implements Signal<T> {
@@ -577,15 +586,6 @@ class _Signal<T> implements Signal<T> {
     if (val != this._value) {
       _updateValue(val);
     }
-  }
-
-  /// Should only be called to update the value of a signal if checks for equality
-  /// have already been made.
-  ///
-  /// This is primarily used by the `ValueSignal` class to update the value
-  /// since the reference is always the same.
-  void forceUpdate(T val) {
-    _updateValue(val);
   }
 
   /// Update the value based on the current value